# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    ArtifactDependentsDto:
      type: object
      properties:
        ecosystem:
          type: string
        name:
          type: string
        latest:
          type: string
        dependents:
          type: array
          items:
            $ref: "#/components/schemas/DependentDto"
...
    DependentDto:
      type: object
      properties:
        artifactEcosystem:
          type: string
        artifactName:
          type: string
        artifactVersion:
          type: string
        repository:
          type: string
        embeddedVersion:
          type: string
        direct:
          type: boolean
        occurredAt:
          $ref: "#/components/schemas/Instant"
        sbomStatus:
          type: string
...
    DownstreamDto:
      type: object
      properties:
        repository:
          type: string
        catalogId:
          type: string
        downstream:
          type: array
          items:
            $ref: "#/components/schemas/EntryDto"
    EntryDto:
      type: object
      properties:
        repository:
          type: string
        catalogId:
          type: string
        archetype:
          type: string
        depth:
          type: integer
          format: int32
        via:
          type: array
          items:
            type: string
    EntryDto1:
      type: object
      properties:
        project:
          type: string
        repository:
          type: string
        ecosystem:
          type: string
        name:
          type: string
        version:
          type: string
        reason:
          type: string
...
    GroupDto:
      type: object
      properties:
        name:
          type: string
        source:
          type: string
        kind:
          type: string
        branch:
          type: string
        state:
          type: string
        headSha:
          type: string
        pending:
          type: integer
          format: int32
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    PinDto:
      type: object
      properties:
        manifestPath:
          type: string
        ecosystem:
          type: string
        name:
          type: string
        version:
          type: string
        range:
          type: string
        kind:
          type: string
        latest:
          type: string
        latestError:
          type: string
        pending:
          type: boolean
        group:
          type: string
        location:
          type: string
        scope:
          type: string
...
    RepositoryDependentsDto:
      type: object
      properties:
        repository:
          type: string
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactDependentsDto"
    RepositoryDetailDto:
      type: object
      properties:
        name:
          type: string
        project:
          type: string
        archetype:
          type: string
        lastScanAt:
          $ref: "#/components/schemas/Instant"
        headSha:
          type: string
        status:
          type: string
        message:
          type: string
        pending:
          type: integer
          format: int32
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupDto"
        pins:
          type: array
          items:
            $ref: "#/components/schemas/PinDto"
        transitives:
          type: array
          items:
            $ref: "#/components/schemas/TransitiveDto"
    RepositoryDto:
      type: object
      properties:
        name:
          type: string
        project:
          type: string
        archetype:
          type: string
        lastScanAt:
          $ref: "#/components/schemas/Instant"
        headSha:
          type: string
        status:
          type: string
        message:
          type: string
        pending:
          type: integer
          format: int32
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupDto"
...
    SbomCheckReportDto:
      type: object
      properties:
        ranAt:
          $ref: "#/components/schemas/Instant"
        filed:
          type: boolean
        entries:
          type: array
          items:
            $ref: "#/components/schemas/EntryDto1"
        warnings:
          type: array
          items:
            type: string
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/TicketDto"
    ScanDto:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        scope:
          type: string
        repository:
          type: string
        trigger:
          type: string
        status:
          type: string
        startedAt:
          $ref: "#/components/schemas/Instant"
        finishedAt:
          $ref: "#/components/schemas/Instant"
        message:
          type: string
...
    StartScanRequest:
      type: object
      properties:
        scope:
          type: string
        repository:
          type: string
    TicketDto:
      type: object
      properties:
        project:
          type: string
        ecosystem:
          type: string
        name:
          type: string
        ticketSlug:
          type: string
        versions:
          type: array
          items:
            type: string
    TransitiveDto:
      type: object
      properties:
        ecosystem:
          type: string
        name:
          type: string
        version:
          type: string
        via:
          type: string
        behind:
          type: boolean
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
...
  /maintenance/api/repositories:
    get:
      summary: "Every repository in the inventory, with its groups and what is pending"
      operationId: listRepositories
      responses:
        "200":
          description: The repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepositoryDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Repository Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/repositories/{name}:
    get:
      summary: One repository with every pin it holds
      operationId: getRepository
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryDetailDto"
        "404":
          description: No such repository in the inventory
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Repository Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/repositories/{name}/dependents:
    get:
      summary: Who embeds the artifacts this repository publishes
      operationId: listRepositoryDependents
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "The dependents, per published artifact"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryDependentsDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Repository Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/repositories/{name}/downstream:
    get:
      summary: "Everything downstream of this repository, ordered upstream first"
      operationId: getRepositoryDownstream
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "The closure, depth ascending then name"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownstreamDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Repository Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/repositories/{name}/groups/{group}/bumps:
    post:
      summary: Put this group's pending changes on its maintenance branch
      operationId: requestGroupBump
      parameters:
      - name: group
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "202":
          description: "Requested; poll GET /bumps/{id}"
        "404":
          description: "No such repository, or no such group"
        "409":
          description: "One is already active, or bumping is disabled"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Repository Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/sbom-check:
    get:
      summary: The last SBOM check's report
      operationId: getSbomCheckReport
      responses:
        "200":
          description: The report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SbomCheckReportDto"
        "404":
          description: No check has run yet
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Sbom Check Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/sbom-check/runs:
    post:
      summary: Run the SBOM check now
      operationId: runSbomCheck
      responses:
        "202":
          description: Ran; the body is the report it stored
        "502":
          description: qits-artifacts could not be asked; nothing stored
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Sbom Check Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
  /maintenance/api/scans:
    post:
      summary: Re-scan the manifests and refresh the latest versions
      operationId: requestScan
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartScanRequest"
        required: true
      responses:
        "202":
          description: Queued
        "400":
          description: Unknown scope
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Scan Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
  /maintenance/api/scans/{id}:
    get:
      summary: One scan and its status
      operationId: getScan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDto"
        "404":
          description: No such scan
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Scan Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
info:
  title: qits-platform-maintenance API
  version: 2026.1002.201448